[map-component-with-geocoder] better structure
